changed int to unsigned long int

+++ apps/kmeans/kernel/kmeans.c
@@ -5,7 +5,7 @@
 
 void assignPointsToClusters(const int64_t *points,  const int64_t *centers, int64_t *clusters,unsigned long int num_points, unsigned long int dimension) {
     //printf("Assign Points to Clusters\n");
-    for (int i = 0; i < num_points; i++) {
+    for (unsigned long int i = 0; i < num_points; i++) {
         int64_t minDist =INT64_MAX;
         for (int c = 0; c < NUM_CLUSTERS; c++) {
             int64_t sum = 0;
@@ -28,23 +28,23 @@ void updateClusterCenters(const int64_t *points, int64_t *centers, int64_t *clus
     int64_t sum[NUM_CLUSTERS][dimension];
     int numbersInCluster[NUM_CLUSTERS];
 
-    for (int i = 0; i < NUM_CLUSTERS; i++) {
+    for (unsigned long int i = 0; i < NUM_CLUSTERS; i++) {
         for (int j = 0; j < dimension; j++) {
             sum[i][j] = 0;
         }
     }
 
-      for (int i = 0; i < NUM_CLUSTERS; i++) {
+      for (unsigned long int i = 0; i < NUM_CLUSTERS; i++) {
             numbersInCluster[i] = 0;
     }
 
 
 
     // Accumulate sum of points for each cluster and count numbers in each cluster
-    for (int j = 0; j < num_points; j++) {
+    for (unsigned long int j = 0; j < num_points; j++) {
         int cluster = clusters[j];
         if (cluster >= 0 && cluster < NUM_CLUSTERS) {
-            for (int d = 0; d < dimension; d++) {
+            for (unsigned long int d = 0; d < dimension; d++) {
                 // Corrected indexing for accessing j-th point's d-th dimension
                 sum[cluster][d] += points[d * num_points + j];
             }
@@ -57,7 +57,7 @@ void updateClusterCenters(const int64_t *points, int64_t *centers, int64_t *clus
         
         if (numbersInCluster[i] > 0) {
 
-            for (int d = 0; d < dimension; d++) {
+            for (unsigned long int d = 0; d < dimension; d++) {
                 // Cast to int64_t if necessary, or consider changing the type of centers to double
                 centers[ d * NUM_CLUSTERS + i] = (int64_t)(sum[i][d] / numbersInCluster[i]);
             }
@@ -67,9 +67,9 @@ void updateClusterCenters(const int64_t *points, int64_t *centers, int64_t *clus
     //Print cluster centers
     printf("Cluster Centers:\n");
     
-    for (int i = 0; i < NUM_CLUSTERS; i++) {
+    for (unsigned long int i = 0; i < NUM_CLUSTERS; i++) {
         printf("(", i);
-        for (int d = 0; d < dimension; d++) {
+        for (unsigned long int d = 0; d < dimension; d++) {
             
             printf("%ld", centers[d * NUM_CLUSTERS + i]);
             if (d < dimension - 1) {
@@ -91,12 +91,12 @@ void assessQualityCluster(const int64_t *points,  int64_t *centers, int64_t *clu
         int64_t numPointsInCluster = 0;
 
         // For each data point
-        for (int dataIndex = 0; dataIndex < num_points; dataIndex++) {
+        for (unsigned long int dataIndex = 0; dataIndex < num_points; dataIndex++) {
             // Check if the data point belongs to the current cluster
             if (clusters[dataIndex] == clusterIndex) {
                 // Calculate the squared distance between the data point and the cluster center
                 int64_t squaredDistance = 0;
-                for (int d = 0; d < dimension; d++) {
+                for (unsigned long int d = 0; d < dimension; d++) {
                     // Adjust indexing to match the row-wise layout of points
                     int64_t diff = points[d * num_points + dataIndex] - centers[d * NUM_CLUSTERS + clusterIndex];
                     squaredDistance += diff * diff;
